requirements.txt
langgraph>=0.2.0
langchain-ollama>=0.1.0
langchain-core>=0.2.0
python-dotenv>=1.0.0